+++ Runtime/Api/Enums/ConsentDebugGeography.cs
@@ -1,0 +1,14 @@
+// ReSharper disable CheckNamespace
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace AppodealStack.Cmp
+{
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
+    public enum ConsentDebugGeography
+    {
+        Eea,
+        RegulatedUsState,
+        Other,
+    }
+}